requirements.txt
jupyterlab
notebook
scipy
matplotlib
kiteconnect
python-dotenv
python-dateutil